requirements.txt
aes-everywhere==1.2.10
alembic==1.7.6
amqp==2.5.2
anyio==3.5.0
# apturl==0.5.2
asgiref==3.4.1
astroid==2.3.3
attrs==20.3.0
autopep8==1.5
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
boto3==1.12.3
botocore==1.15.3
# Brlapi==0.8.2
cached-property==1.5.2
celery==4.4.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
clamd==1.0.2
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==2.9.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.4
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==4.0
django-axes==5.8.0
django-celery-beat==2.0.0
django-clamd==0.4.0
django-cors-headers==3.2.1
django-dirtyfields==1.4
django-environ==0.4.5
django-extensions==2.2.8
django-fernet-fields==0.6
django-filter==2.2.0
django-import-export==2.0.2
django-ipware==3.0.1
django-phonenumber-field==4.0.0
django-ratelimit==3.0.1
django-storages==1.9.1
django-timezone-field==4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docker==4.4.4
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.20
ecdsa==0.17.0
et-xmlfile==1.0.1
fastapi==0.74.1
fastapi-utils==0.2.1
fasteners==0.14.1
fcm-django==0.3.4
filelock==3.6.0
future==0.18.2
greenlet==1.1.2
gTTS==2.2.3
h11==0.13.0
h2==2.6.2
hpack==3.0.0
httplib2==0.18.1
httptools==0.3.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jdcal==1.4.1
jeepney==0.7.1
jmespath==0.9.4
joblib==1.1.0
jsonschema==3.2.0
keyring==23.0.1
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
netifaces==0.10.9
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
paramiko==2.7.2
passlib==1.7.4
pexpect==4.8.0
phonenumbers==8.11.4
Pillow==8.1.2
pipenv==2018.11.26
platformdirs==2.5.1
pluginmanager==0.4.1
protobuf==3.12.4
pushjack==1.6.0
pyasn1==0.4.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
pycups==2.0.1
pydantic==1.9.0
pydotplus==2.0.2
pydub==0.25.1
pyfcm==1.4.7
PyGObject==3.40.1
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==3.0.0b1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.2.1
python-crontab==2.4.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.19.2
python-jose==3.3.0
python-magic==0.4.15
python-multipart==0.0.5
pyttsx3==2.90
pytz==2019.3
pyxdg==0.27
PyYAML==5.3.1
qrcode==7.3
razorpay==1.2.0
redis==3.4.1
regex==2021.11.10
reportlab==3.5.45
requests==2.23.0
rsa==4.8
s3transfer==0.3.3
SecretStorage==3.3.1
simplejson==3.17.2
six==1.14.0
sniffio==1.2.0
SpeechRecognition==3.8.1
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
starlette==0.17.1
systemd-python==234
tablib==1.1.0
texttable==1.6.3
tqdm==4.62.3
twilio==6.42.0
typed-ast==1.4.1
typing_extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.17.5
uvloop==0.16.0
vine==1.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
wadllib==1.3.5
watchgod==0.7
websocket-client==0.57.0
websockets==10.2
wrapt==1.11.2
xdg==5
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
